requirements.txt
torch==2.10.0
torchvision==0.25.0
numpy
pandas
albumentations
scikit-learn
Pillow
flask